sf0x: split into separate header, implementation, and main
