+++ src/drivers/distance_sensor/sf0x/CMakeLists.txt
@@ -36,7 +36,9 @@ px4_add_module(
 	COMPILE_FLAGS
 		-Wno-cast-align # TODO: fix and enable
 	SRCS
-		sf0x.cpp
+		SF0X.cpp
+		SF0X.hpp
+		sf0x_main.cpp
 		sf0x_parser.cpp
 	MODULE_CONFIG
 		module.yaml
